Corrected numerical cast

+++ Source/Urho3D/UI/Slider.cpp
@@ -256,7 +256,7 @@ void Slider::UpdateSlider()
         else
         {
             const int sliderLength{knob_->IsFixedHeight() ? knob_->GetHeight()
-                                                          : Max(static_cast<float>(GetHeight() / (range_ + 1.0f)),
+                                                          : Max(static_cast<int>(GetHeight() / (range_ + 1.0f)),
                                                                 border.top_ + border.bottom_)};
 
             const float sliderPos{(GetHeight() - sliderLength) * value_ / range_};
